backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "book-character-simulation-backend"
version = "0.1.0"
description = "Backend services for the book character simulation project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Flask>=3.0,<4.0",
  "Flask-Cors>=4.0,<6.0",
  "PyMuPDF>=1.24,<2.0",
  "chromadb>=0.6,<0.7",
  "google-genai>=0.5,<1.0",
  "openai>=1.35,<2.0",
  "pydantic>=2.8,<3.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "gunicorn>=22.0,<24.0",
  "pytest>=8.0,<9.0",
  "ruff>=0.6,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
